+++ backend/cheq/src/main/java/potato/cheq/dto/request/RequestAdminDto.java
@@ -1,4 +1,4 @@
-package potato.cheq.dto;
+package potato.cheq.dto.request;
 
 public class RequestAdminDto {
 }
