+++ backend/cheq/src/main/java/potato/cheq/dto/request/RequestLoginDto.java
@@ -1,5 +1,4 @@
-package potato.cheq.dto;
-
+package potato.cheq.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,7 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RequestUpdateUserDto {
-    private String uuid;
+public class RequestLoginDto {
+    private String studentId;
 }
-
